# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SistemaJava • Mercadinho PDV",
    "description": "Obtenha CSRF em /api/v1/auth/csrf; envie X-XSRF-TOKEN em mutações. Login retorna Bearer JWT.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/api/v1/stock-movements": {
      "get": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "movements",
        "parameters": [
          {
            "name": "productId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageStockMovement"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "catalog-controller"
        ],
        "operationId": "stock",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StockMovement"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cash/sessions/{id}/movements": {
      "post": {
        "tags": [
          "cash-controller"
        ],
        "operationId": "movement",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashMovementInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Summary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cash/sessions/{id}/close": {
      "post": {
        "tags": [
          "cash-controller"
        ],
        "operationId": "close",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashClose"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Summary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cash/open": {
      "post": {
        "tags": [
          "cash-controller"
        ],
        "operationId": "open",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashOpen"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Summary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/backup/restore": {
      "post": {
        "tags": [
          "backup-controller"
        ],
        "operationId": "restore",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/writeoffs/summary": {
      "get": {
        "tags": [
          "stock-controller"
        ],
        "operationId": "summary",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cash/sessions": {
      "get": {
        "tags": [
          "cash-controller"
        ],
        "operationId": "list_3",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageCashSession"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cash/sessions/{id}": {
      "get": {
        "tags": [
          "cash-controller"
        ],
        "operationId": "summary_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Summary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Customer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "document": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "creditLimit": {
            "type": "number"
          },
          "balance": {
            "type": "number"
          },
          "address": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "termDays": {
            "type": "integer",
            "format": "int32"
          },
          "interestDay": {
            "type": "number"
          },
          "penaltyDay": {
            "type": "number"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "markup": {
            "type": "number"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
      "StockInput": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string",
            "pattern": "IN|OUT|ADJUSTMENT"
          },
          "quantity": {
            "type": "number",
            "minimum": 0
          },
          "reason": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "expiresOn": {
            "type": "string",
            "format": "date"
          },
          "unitCost": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "productId",
          "quantity",
          "reason",
          "type"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "barcode": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "unit": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "cost": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "minimumStock": {
            "type": "number"
          },
          "expiresOn": {
            "type": "string",
            "format": "date"
          },
          "perishable": {
            "type": "boolean"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
      "Sale": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "requestHash": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "total": {
            "type": "number"
          },
          "fees": {
            "type": "number"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "StockMovement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "sale": {
            "$ref": "#/components/schemas/Sale"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          }
        }
      },
...
      "CashMovementInput": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "pattern": "WITHDRAWAL|SUPPLY"
          },
          "amount": {
            "type": "number",
            "minimum": 0
          },
          "reason": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "amount",
          "reason",
          "type"
        ]
      },
      "CashMovement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "type": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "reason": {
            "type": "string"
          }
        }
      },
      "CashSession": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "openedAt": {
            "type": "string",
            "format": "date-time"
          },
          "openingAmount": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "closedAt": {
            "type": "string",
            "format": "date-time"
          },
          "closedBy": {
            "$ref": "#/components/schemas/User"
          },
          "expectedCash": {
            "type": "number"
          },
          "countedCash": {
            "type": "number"
          },
          "difference": {
            "type": "number"
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "PaymentTotal": {
        "type": "object",
        "properties": {
          "method": {
            "type": "string"
          },
          "total": {
            "type": "number"
          }
        }
      },
      "Summary": {
        "type": "object",
        "properties": {
          "session": {
            "$ref": "#/components/schemas/CashSession"
          },
          "sales": {
            "type": "integer",
            "format": "int64"
          },
          "cancelled": {
            "type": "integer",
            "format": "int64"
          },
          "salesTotal": {
            "type": "number"
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentTotal"
            }
          },
          "cashSales": {
            "type": "number"
          },
          "creditReceipts": {
            "type": "number"
          },
          "supplies": {
            "type": "number"
          },
          "withdrawals": {
            "type": "number"
          },
          "expectedCash": {
            "type": "number"
          },
          "movements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CashMovement"
            }
          }
        }
      },
      "CashClose": {
        "type": "object",
        "properties": {
          "countedCash": {
            "type": "number",
            "minimum": 0
          },
          "notes": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          }
        },
        "required": [
          "countedCash"
        ]
      },
      "CashOpen": {
        "type": "object",
        "properties": {
          "openingAmount": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "openingAmount"
        ]
      },
      "RestoreInput": {
        "type": "object",
        "properties": {
          "filename": {
            "type": "string",
            "minLength": 1,
            "pattern": "backup-[0-9TZ-]+[.]dump"
          },
          "confirmation": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "confirmation",
          "filename"
        ]
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "sorted": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
      "PageStockMovement": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "last": {
            "type": "boolean"
          },
          "first": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StockMovement"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "PageCashSession": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "last": {
            "type": "boolean"
          },
          "first": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CashSession"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}